requirements.txt
Django>=1.8,<1.9
RapidSMS==0.21.0
gunicorn==0.17.4
whitenoise==3.3.1
dj-database-url==0.4.2
psycopg2==2.7.1
rapidsms-twilio==1.0.1
mock==2.0.0
django-address==0.1.5
